# documentation/openapi/providers.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "providers API",
    "description": "Internal HTTP API for the providers service.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/internal/quotes": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "symbols"
                ],
                "properties": {
                  "symbols": {
                    "minItems": 1,
                    "maxItems": 200,
                    "type": "array",
                    "items": {
                      "minLength": 1,
                      "maxLength": 40,
                      "type": "string"
                    }
                  },
                  "provider": {
                    "minLength": 1,
                    "maxLength": 64,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "provider",
                    "quotes"
                  ],
                  "properties": {
                    "provider": {
                      "type": "string"
                    },
                    "quotes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "symbol",
                          "price",
                          "previousClose",
                          "currency",
                          "timestampMs"
                        ],
                        "properties": {
                          "symbol": {
                            "type": "string"
                          },
                          "price": {
                            "type": "string"
                          },
                          "previousClose": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "currency": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "timestampMs": {
                            "anyOf": [
                              {
                                "type": "integer"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "series": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "required": [
                                "timeMs",
                                "close"
                              ],
                              "properties": {
                                "timeMs": {
                                  "type": "integer"
                                },
                                "close": {
                                  "type": "string"
                                },
                                "volume": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ]
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/earnings": {
      "get": {
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 40,
              "type": "string"
            },
            "in": "query",
            "name": "symbol",
            "required": true
          },
          {
            "schema": {
              "minLength": 1,
              "maxLength": 64,
              "type": "string"
            },
            "in": "query",
            "name": "provider",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "provider",
                    "earnings"
                  ],
                  "properties": {
                    "provider": {
                      "type": "string"
                    },
                    "earnings": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}